// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** A date string, such as 2007-12-03, compliant with the `full-date` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any;
};
// ...
export type Page = Document & {
  __typename?: 'Page';
  /** Document ID */
  _id?: Maybe<Scalars['ID']>;
  /** Document type */
  _type?: Maybe<Scalars['String']>;
  /** Date the document was created */
  _createdAt?: Maybe<Scalars['DateTime']>;
  /** Date the document was last modified */
  _updatedAt?: Maybe<Scalars['DateTime']>;
  /** Current document revision */
  _rev?: Maybe<Scalars['String']>;
  _key?: Maybe<Scalars['String']>;
  /** Page <title> */
  title?: Maybe<Scalars['String']>;
  /** Should this page be available on the site? */
  active?: Maybe<Scalars['Boolean']>;
  /** Pathname the page should be accessible at e.g. https://roadtofree.md/england */
  pathname?: Maybe<Scalars['String']>;
  /** Content for <meta name="description"> */
  description?: Maybe<Scalars['String']>;
  /** Content for <link rel="canonical"> */
  canonical?: Maybe<Scalars['String']>;
  /** Name of the country used in links */
  countryName?: Maybe<Scalars['String']>;
  /** This emoji will be shown next to the country name */
  emoji?: Maybe<Scalars['String']>;
  /** Content for <h1> */
  headingOne?: Maybe<Scalars['String']>;
  /** Content for the description paragraph shown at the top of the page */
  pageDescription?: Maybe<Scalars['String']>;
  timeline?: Maybe<Array<Maybe<Step>>>;
  /** Show this page content on the home page? */
  indexPage?: Maybe<Scalars['Boolean']>;
  /** Content for <title> when this page is being used on the home page */
  indexTitle?: Maybe<Scalars['String']>;
  /** Content for <meta name="description"> when this page is being used on the home page */
  indexDescription?: Maybe<Scalars['String']>;
  /** Content for <h1> when this page is being used on the home page */
  indexHeadingOne?: Maybe<Scalars['String']>;
  /** Content for the description paragraph shown at the top of the page when this page is being used on the home page */
  indexPageDescription?: Maybe<Scalars['String']>;
};

/** A Sanity document */
export type Document = {
  /** Document ID */
  _id?: Maybe<Scalars['ID']>;
  /** Document type */
  _type?: Maybe<Scalars['String']>;
  /** Date the document was created */
  _createdAt?: Maybe<Scalars['DateTime']>;
  /** Date the document was last modified */
  _updatedAt?: Maybe<Scalars['DateTime']>;
  /** Current document revision */
  _rev?: Maybe<Scalars['String']>;
};

export type Step = Document & {
  __typename?: 'Step';
  /** Document ID */
  _id?: Maybe<Scalars['ID']>;
  /** Document type */
  _type?: Maybe<Scalars['String']>;
  /** Date the document was created */
  _createdAt?: Maybe<Scalars['DateTime']>;
  /** Date the document was last modified */
  _updatedAt?: Maybe<Scalars['DateTime']>;
  /** Current document revision */
  _rev?: Maybe<Scalars['String']>;
  _key?: Maybe<Scalars['String']>;
  /** Only used internally to help you recognise this timeline step */
  title?: Maybe<Scalars['String']>;
  /** The date this timeline step comes into effect. Use https://timestampgenerator.com to generate the correct format. Make sure you copy the "ISO 8601" format. */
  date?: Maybe<Scalars['String']>;
  activities?: Maybe<Array<Maybe<Activity>>>;
};

export type Activity = Document & {
  __typename?: 'Activity';
  /** Document ID */
  _id?: Maybe<Scalars['ID']>;
  /** Document type */
  _type?: Maybe<Scalars['String']>;
  /** Date the document was created */
  _createdAt?: Maybe<Scalars['DateTime']>;
  /** Date the document was last modified */
  _updatedAt?: Maybe<Scalars['DateTime']>;
  /** Current document revision */
  _rev?: Maybe<Scalars['String']>;
  _key?: Maybe<Scalars['String']>;
  /** Only used internally to help you recognise this activity */
  title?: Maybe<Scalars['String']>;
  /** This content is shown to the user */
  description?: Maybe<Scalars['String']>;
  /** This emoji is shown next to the description */
  emoji?: Maybe<Scalars['String']>;
  categories?: Maybe<Array<Maybe<Category>>>;
};

export type Category = Document & {
  __typename?: 'Category';
  /** Document ID */
  _id?: Maybe<Scalars['ID']>;
  /** Document type */
  _type?: Maybe<Scalars['String']>;
  /** Date the document was created */
  _createdAt?: Maybe<Scalars['DateTime']>;
  /** Date the document was last modified */
  _updatedAt?: Maybe<Scalars['DateTime']>;
  /** Current document revision */
  _rev?: Maybe<Scalars['String']>;
  _key?: Maybe<Scalars['String']>;
  /** Only used internally to help you recognise this category */
  title?: Maybe<Scalars['String']>;
  /** Label for category */
  label?: Maybe<Scalars['String']>;
  /** Emoji shown next to the label */
  emoji?: Maybe<Scalars['String']>;
};